에너미, 보스 조건 반영
스테이지3 에너미 active 조건 반영.
보스 전투 조건 및 2페이즈 죽음시 퀘스트 카운팅 반영.

+++ Assets/00.JY/Script/EnemyHealth.cs
@@ -74,9 +74,11 @@ public class EnemyHealth : MonoBehaviour
                 ani.SetTrigger("IsDie");
             isdie = true;
             agent.enabled = false;
-            PlayerInteraction.instance.Kill();
+            if (!QuestManager.instance.bossAction)
+                PlayerInteraction.instance.Kill();
             UIManager.instance.UpdateExp(exp);
-            Instantiate(Resources.Load<GameObject>("Prefabs/DropItem"), this.transform.position, Quaternion.identity);
+            if (!QuestManager.instance.bossAction)
+                Instantiate(Resources.Load<GameObject>("Prefabs/DropItem"), this.transform.position, Quaternion.identity);
             rigid.Sleep();
             capcol.enabled = false;
             HpBarCanvas.SetActive(false);
